requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
dm-tree==0.1.9
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
more-itertools==10.7.0
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-model-optimization==0.8.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
